Move getGlobal into separate file to remove circular dep
