+++ packages/util/index.node.ts
@@ -42,3 +42,4 @@ export * from './src/uuid';
 export * from './src/exponential_backoff';
 export * from './src/formatters';
 export * from './src/compat';
+export * from './src/global';
