+++ packages/util/index.ts
@@ -37,3 +37,4 @@ export * from './src/uuid';
 export * from './src/exponential_backoff';
 export * from './src/formatters';
 export * from './src/compat';
+export * from './src/global';
